Refactor category deletion logic and add confirmation dialog

+++ client/src/page/admin/categories/list.jsx
@@ -1,18 +1,35 @@
 import React, { useCallback, useEffect, useState } from "react";
 import useCategory from "../../../hooks/useCategory";
+import Swal from "sweetalert2";
+
 import { Link } from "react-router-dom";
 
 export default function list() {
   const { list, deleteById } = useCategory();
   const [cats, setCats] = useState([]);
 
-  const handleDelete = async (e, id) => {
-    e.preventDefault();
+  const handleDelete = async (event, id) => {
+    event.preventDefault();
     try {
-      const resp = await deleteById(id);
-    
-      if (resp) {
-        await list();
+      const result = await Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!",
+      });
+      if (result.isConfirmed) {
+        const resp = await deleteById(id);
+        if (resp) {
+          Swal.fire({
+            title: "Deleted!",
+            text: "Delete Successful.",
+            icon: "success",
+          });
+          await list();
+        }
       }
     } catch (err) {
       alert(err || "Something went wrong");
@@ -21,17 +38,19 @@ export default function list() {
 
   const fetchCat = useCallback(async () => {
     const result = await list();
-    setCats(result.data);
-    return result;
-  }, []);
+    if(result){
+      setCats(result.data);
+    }
+ 
+  }, [list]);
 
   useEffect(() => {
     fetchCat();
   }, [fetchCat]);
-console.log(cats)
+
   return (
     <div className="overflow-x-auto">
-      <Link to={'/admin/categories/add'}>Add Category</Link>
+      <Link to={"/admin/categories/add"}>Add Category</Link>
       <table className="min-w-full bg-white border border-gray-300">
         <thead>
           <tr>
@@ -52,7 +71,7 @@ console.log(cats)
                 </button>
               </td>
               <td className="py-2 px-4 border-b">
-                <Link to={`/admin/categories/${cat?._id}`}></Link>
+                <Link to={`/admin/categories/${cat?._id}`}>Edit</Link>
               </td>
             </tr>
           ))}
